+++ javascripts/main.js
@@ -1,3 +1,5 @@
 "use strict";
 
 require("./foodie-list");
+require("./city-dropdown");
+require("./change-city");
