+++ sass/main.scss
@@ -21,8 +21,11 @@
 }
 
 .reason{
-    font-size: 1em;
-    
+    font-size: 1em; 
+}
+
+#emptySelection{
+   padding: 50px;
 }
 
 @media only screen and (max-width: 500px) {
